Use new variables in build file

+++ jenkins-scripts/_bitbucket_create_build_status_file.bash
@@ -6,7 +6,6 @@ SCRIPT_DIR="${SCRIPT_DIR%/*}"
 
 echo '# BEGIN SECTION: sending bitbucket status: inprogress'
 NEEDED_HOST_PACKAGES="python-pip"
-REPO_SHORT_NAME=`echo ${SRC_REPO} | sed s:.*\.org/::`
 
 # Source bitbucket configs
 . ${SCRIPT_DIR}/_bitbucket_configs.bash
@@ -16,10 +15,23 @@ if [[ ! -f ${BITBUCKET_BUILD_STATUS_FILE} ]]; then
   exit 1
 fi
 
+if [[ ! -f ${JENKINS_BUILD_HG_HASH} ]]; then
+  echo "JENKINS_BUILD_HG_HASH variable missing"
+  exit 1
+fi
+
+if [[ ! -f ${JENKINS_BUILD_REPO} ]]; then
+  echo "JENKINS_BUILD_REPO variable missing"
+  exit 1
+fi
+
+
 if [[ ! -f ${BITBUCKET_USER_PASS_FILE} ]]; then
   echo "Bitbucket user pass not found in file \${BITBUCKET_USER_PASS_FILE}"
   exit 1
 fi
+
+REPO_SHORT_NAME=`echo ${JENKINS_BUILD_REPO} | sed s:.*\.org/::`
 
 echo "Generating requirements.txt for pip"
 cat > ${WORKSPACE}/requirements.txt <<- REQ
@@ -48,7 +60,7 @@ echo "Generating config file ..."
 cat > $BITBUCKET_BUILD_STATUS_FILE << DELIM_CONFIG
 bitbucket_origin:
   repository_name: ${REPO_SHORT_NAME}
-  sha: ${MERCURIAL_REVISION_SHORT}
+  sha: ${JENKINS_BUILD_HG_HASH}
 jenkins_job:
   name: ${JENKINS_BUILD_JOB_NAME}
   url: ${JENKINS_BUILD_URL}
